Move font URL input outside graph
