+++ src/ReactFlow.jsx
@@ -6,7 +6,6 @@ import {
 } from "@xyflow/react";
 import { Resizable } from "re-resizable";
 import ErrorBoundary from "@/components/ErrorBoundary.jsx";
-import { Input } from "@/components/ui/input.jsx";
 import "@xyflow/react/dist/style.css";
 
 /**
@@ -34,39 +33,6 @@ const ReactFlow = ({
 }) => {
   const [zoom, setZoom] = useState(0.7);
   const [focus, setFocus] = useState({ x: 0, y: 0 });
-  const [fontUrl, setFontUrl] = useState("");
-
-  const handleFontUrlChange = (e) => {
-    const url = e.target.value;
-    setFontUrl(url);
-
-    const existing = document.getElementById("dynamic-node-font");
-    if (existing) existing.remove();
-
-    if (url) {
-      const link = document.createElement("link");
-      link.id = "dynamic-node-font";
-      link.rel = "stylesheet";
-      link.href = url;
-      document.head.appendChild(link);
-
-      try {
-        const u = new URL(url);
-        const familyParam = u.searchParams.get("family");
-        if (familyParam) {
-          const family = decodeURIComponent(familyParam.split(":")[0]).replace(/\+/g, " ");
-          document.documentElement.style.setProperty(
-            "--node-font-family",
-            `'${family}', sans-serif`
-          );
-        }
-      } catch {
-        // ignore URL parse errors
-      }
-    } else {
-      document.documentElement.style.removeProperty("--node-font-family");
-    }
-  };
 
   return (
     <Resizable
@@ -83,15 +49,6 @@ const ReactFlow = ({
         Zoom: {zoom.toFixed(2)} • {Math.round(rfSize.width)}×
         {Math.round(rfSize.height)} • Focus: {Math.round(focus.x)},
         {Math.round(focus.y)}
-      </div>
-      <div className="absolute top-1 right-1 z-10 w-48">
-        <Input
-          type="text"
-          placeholder="Font CSS URL"
-          value={fontUrl}
-          onChange={handleFontUrlChange}
-          className="h-8"
-        />
       </div>
       <div
         className="w-full h-full"
